// src/app/styled-css/margin.style/margin.d.ts
import { Property } from 'csstype';

export interface MarginStyleProps {
  margin?: Property.Margin;
  marginXxl?: Property.Margin;
  marginXl?: Property.Margin;
  marginLg?: Property.Margin;
  marginMd?: Property.Margin;
  marginSm?: Property.Margin;
  marginXs?: Property.Margin;

  marginTop?: Property.Margin;
  marginTopXxl?: Property.Margin;
  marginTopXl?: Property.Margin;
  marginTopLg?: Property.Margin;
  marginTopMd?: Property.Margin;
  marginTopSm?: Property.Margin;
  marginTopXs?: Property.Margin;

  marginBottom?: Property.Margin;
  marginBottomXxl?: Property.Margin;
  marginBottomXl?: Property.Margin;
  marginBottomLg?: Property.Margin;
  marginBottomMd?: Property.Margin;
  marginBottomSm?: Property.Margin;
  marginBottomXs?: Property.Margin;

  marginRight?: Property.Margin;
  marginRightXxl?: Property.Margin;
  marginRightXl?: Property.Margin;
  marginRightLg?: Property.Margin;
  marginRightMd?: Property.Margin;
  marginRightSm?: Property.Margin;
  marginRightXs?: Property.Margin;

  marginLeft?: Property.Margin;
  marginLeftXxl?: Property.Margin;
  marginLeftXl?: Property.Margin;
  marginLeftLg?: Property.Margin;
  marginLeftMd?: Property.Margin;
  marginLeftSm?: Property.Margin;
  marginLeftXs?: Property.Margin;

  marginTopBottom?: Property.Margin;
  marginTopBottomXxl?: Property.Margin;
  marginTopBottomXl?: Property.Margin;
  marginTopBottomLg?: Property.Margin;
  marginTopBottomMd?: Property.Margin;
  marginTopBottomSm?: Property.Margin;
  marginTopBottomXs?: Property.Margin;

  marginLeftRight?: Property.Margin;
  marginLeftRightXxl?: Property.Margin;
  marginLeftRightXl?: Property.Margin;
  marginLeftRightLg?: Property.Margin;
  marginLeftRightMd?: Property.Margin;
  marginLeftRightSm?: Property.Margin;
  marginLeftRightXs?: Property.Margin;
}